\name{closeAllNcfiles}
\alias{closeAllNcfiles}
\title{Close all open RnetCDF file connections}
\description{closeAllNcfiles is a convenience function to close all netCDF connections that are 
currently open. }
\usage{closeAllNcfiles()}

\value{nothing is returned.}

\author{Jannis v. Buttlar}





