\name{modifyNcdfStdFile}
\alias{modifyNcdfStdFile}
\title{Standardize file name and missing value attribute of a ncdf file.}
\description{Wrapper function around modifyNcdfStdFile and modifyNcdfSetMissing
to standardize the name and the missing value attributes of a ncdf file.}
\usage{modifyNcdfStdFile(file.con)}
\arguments{
  \item{file.con}{character string (i.e. file name) or file connection of the file to change.}
}



\author{Jannis v. Buttlar}





