\name{modifyNcdfStdNames}
\alias{modifyNcdfStdNames}
\title{modify non standard longitude and latitude names}
\description{This function modifies lat, lon and long dimension names to 'latitude'
and 'longitude'}
\usage{modifyNcdfStdNames(fileCon)}\arguments{
  \item{fileCon}{file in which to modify the names. can be supplied as a character vector
or as a ncdf file connection.}
}

\value{Nothing is returned}

\author{Jannis v. Buttlar}





