\name{ncdf.tools-package}
\alias{ncdf.tools-package}
\alias{ncdf.tools}
\docType{package}
\title{Set of tools to simplify the handling of NetCDF files.}
\description{Set of tools to simplify the handling of ncdf files with the RNetCDF package.
Most functions are wrappers for basic functions from RNetCDF to run combinations of these
functions for frequently encountered tasks. A few functions included are very specific
to the authors research work and may be of limited use to others. All code is under
constant development and errors or bugs sometimes occour when the functions are used in
new contexts and scenarios which have not yet been tested. Please report all such errors to
the mailing lists on R-Forge and the author will try to fix them, time permitting.}
\details{
\tabular{ll}{Package: \tab ncdf.tools\cr
Title: \tab Set of tools to simplify the handling of NetCDF files.\cr
Version: \tab 0.7.0.283\cr
Date: \tab 2011-02-15\cr
Author: \tab Jannis v. Buttlar\cr
Maintainer: \tab Jannis v. Buttlar <jbuttlar@bgc-jena.mpg.de>\cr
Imports: \tab RNetCDF, chron, foreach, iterators, codetools, multicore, doMC,
DistributionUtils, abind, plotrix, snow, snowfall, raster, RColorBrewer, plyr,
Rssa, JBTools, doSNOW\cr
License: \tab GPL-2\cr
LazyLoad: \tab yesf\cr
Depends: \tab R (>= 2.10.0)\cr}
}
\author{Jannis v. Buttlar}

\keyword{ package }


