\name{modifyNcdfAppendHistory}
\alias{modifyNcdfAppendHistory}
\title{Append a string to netCDF history}
\description{Convenience function to append a string together with the date and the user
to the history attribute of an NetCDF file. }
\usage{modifyNcdfAppendHistory(file, string)}\arguments{
  \item{file}{character sting or RNetCDF file connection: file to write to.}
  \item{string}{character string: string to append to the history}
}

\value{nothing is returned}

\author{Jannis v. Buttlar}





