\name{modifyNcdfStdNames}
\alias{modifyNcdfStdNames}
\title{Modify non standard longitude and latitude names}
\description{This function modifies dimension names like 'lat', 'lon' and 'long' to 'latitude'
and 'longitude'.}
\usage{modifyNcdfStdNames(fileCon)}
\arguments{
  \item{fileCon}{file in which to modify the names. can be supplied as a character vector
or as a netCDF file connection.}
}

\value{Nothing is returned}

\author{Jannis v. Buttlar}





