\name{ncdf.tools-package}
\alias{ncdf.tools-package}
\alias{ncdf.tools}
\docType{package}
\title{Easier ncdf file handling}
\description{Set of tools to simplify the handling of ncdf files with the RNetCDF package.
Most functions are wrappers for basic functions from RNetCDF to easily run combinations of these
functions for frequently encountered tasks.}
\details{
\tabular{ll}{Package: \tab ncdf.tools\cr
Title: \tab Easier ncdf file handling\cr
Version: \tab 0.7.1.289\cr
Date: \tab 2014-01-01\cr
Author: \tab Jannis v. Buttlar\cr
Maintainer: \tab Jannis v. Buttlar <jbuttlar@bgc-jena.mpg.de>\cr
Imports: \tab RNetCDF, chron, foreach, doMC, abind, plotrix, raster, RColorBrewer, JBTools\cr
License: \tab GPL-2\cr
LazyLoad: \tab yesf\cr
Depends: \tab R (>= 2.10.0)\cr}
}
\author{Jannis v. Buttlar}

\keyword{ package }


