\name{ncdf.tools-package}
\alias{ncdf.tools-package}
\alias{ncdf.tools}
\docType{package}
\title{Easier 'NetCDF' File Handling}
\description{Set of tools to simplify the handling of 'NetCDF' files with the 'RNetCDF' package.
Most functions are wrappers of basic functions from the 'RNetCDF' package to easily run combinations of these
functions for frequently encountered tasks.}
\details{
\tabular{ll}{Package: \tab ncdf.tools\cr
Title: \tab Easier 'NetCDF' File Handling\cr
Version: \tab 0.7.1.295\cr
Date: \tab 2015-05-20\cr
Author: \tab Jannis v. Buttlar\cr
Maintainer: \tab Jannis v. Buttlar <jbuttlar@bgc-jena.mpg.de>\cr
Imports: \tab RNetCDF, chron, parallel, abind, plotrix, raster, RColorBrewer, JBTools\cr
License: \tab GPL-2\cr
LazyLoad: \tab yesf\cr
Depends: \tab R (>= 2.10.0)\cr}
}
\author{Jannis v. Buttlar}

\keyword{ package }


