\name{ncdf.tools-package}
\alias{ncdf.tools-package}
\alias{ncdf.tools}
\docType{package}
\title{Set of tools to ease the handling of ncdf files}
\description{Set of tools to ease the handling of ncdf files}
\details{
\tabular{ll}{Package: \tab ncdf.tools\cr
Title: \tab Set of tools to ease the handling of ncdf files\cr
Version: \tab 0.7.0.281\cr
Date: \tab 2011-02-15\cr
Author: \tab Jannis v. Buttlar\cr
Maintainer: \tab Jannis v. Buttlar <jbuttlar@bgc-jena.mpg.de>\cr
Imports: \tab RNetCDF, chron, foreach, iterators, codetools, multicore, doMC,
DistributionUtils, abind, plotrix, snow, snowfall, raster, RColorBrewer, plyr,
Rssa, JBTools, doSNOW\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
Depends: \tab R (>= 2.10.0)\cr}
}
\author{Jannis v. Buttlar}

\keyword{ package }


